Change all *_elemental_effect from 1-based to 0-based
This does NOT include the data!
That is, res.cpp is still using a mixture of 0- and 1-based values.

+++ include/fighter.h
@@ -332,7 +332,7 @@ class KFighter
      *
      * See eResistance enum.
      */
-    int weapon_elemental_effect;
+    uint8_t weapon_elemental_effect;
 
     /*! \brief UNLiving (undead), like zombies, skeletons, etc. */
     int unl;
